requirements.txt
torch
transformers
numpy
torchvision
pillow
tqdm
scikit-learn
clip
torchsummary
